Share gift

+++ feature-gift-impl/src/main/java/io/novafoundation/nova/feature_gift_impl/presentation/gifts/GiftRVItem.kt
@@ -1,5 +1,5 @@
 package io.novafoundation.nova.feature_gift_impl.presentation.gifts
 
 class GiftRVItem(
-    val id: Int // TODO add otherFields
+    val id: Long // TODO add otherFields
 )
